requirements.txt
blinker==1.9.0
click==8.1.7
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
pillow==11.0.0
Werkzeug==3.1.3
